Reclaim flash
Reclaim a small amount of flash by not bothering to disable IRQs that should never be enabled by bootloader.

There's room to cut a few more of the IRQ disables, but I didn't want to touch anything I wasn't 100% sure about. If double-checked, probably a dozen more could be cut.